+++ hypervisor/boot/include/multiboot2.h
@@ -79,69 +79,4 @@
 #define MULTIBOOT2_ARCHITECTURE_I386			0
 #define MULTIBOOT2_ARCHITECTURE_MIPS32			4
 
-#ifndef ASSEMBLER
-struct multiboot2_mmap_entry
-{
-	uint64_t addr;
-	uint64_t len;
-	uint32_t type;
-	uint32_t zero;
-};
-
-struct multiboot2_tag
-{
-	uint32_t type;
-	uint32_t size;
-};
-
-struct multiboot2_tag_string
-{
-	uint32_t type;
-	uint32_t size;
-	char string[0];
-};
-
-struct multiboot2_tag_module
-{
-	uint32_t type;
-	uint32_t size;
-	uint32_t mod_start;
-	uint32_t mod_end;
-	char cmdline[0];
-};
-
-struct multiboot2_tag_mmap
-{
-	uint32_t type;
-	uint32_t size;
-	uint32_t entry_size;
-	uint32_t entry_version;
-	struct multiboot2_mmap_entry entries[0];
-};
-
-struct multiboot2_tag_new_acpi
-{
-	uint32_t type;
-	uint32_t size;
-	uint8_t rsdp[0];
-};
-
-struct multiboot2_tag_efi64
-{
-	uint32_t type;
-	uint32_t size;
-	uint64_t pointer;
-};
-
-struct multiboot2_tag_efi_mmap
-{
-	uint32_t type;
-	uint32_t size;
-	uint32_t descr_size;
-	uint32_t descr_vers;
-	uint8_t efi_mmap[0];
-};
-
-#endif /* ASSEMBLER */
-
 #endif /*  MULTIBOOT2_H */
